pyproject.toml
[project]
name = "protocol-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "jinja2>=3.0.0",
    "python-dotenv>=0.19.0",
    "sqlalchemy>=1.4.0",
]
